+++ src/test/java/com/wicio/shiplog/route/DistanceCalculatorBasedOnSpeedAndTimeTest.java
@@ -24,7 +24,7 @@ class DistanceCalculatorBasedOnSpeedAndTimeTest {
     @ParameterizedTest
     @MethodSource("provideTestValuesForDistanceInMetres")
     void shouldCalculateDistanceInMetres(double expected,
-                                         Double speedOverGroundInKmPerH,
+                                         Integer speedOverGroundInKmPerH,
                                          long minutesBetween) {
       //given
       //when
@@ -35,21 +35,21 @@ class DistanceCalculatorBasedOnSpeedAndTimeTest {
 
     @Test
     void shouldThrowErrorWhenSpeedIsNegative() {
-      assertThrows(IllegalStateException.class, () -> testObj.distanceInMetres(-1.0, 1));
+      assertThrows(IllegalStateException.class, () -> testObj.distanceInMetres(-1, 1));
     }
 
     @Test
     void shouldThrowErrorWhenTimeIsNegative() {
-      assertThrows(IllegalStateException.class, () -> testObj.distanceInMetres(1.0, -1));
+      assertThrows(IllegalStateException.class, () -> testObj.distanceInMetres(1, -1));
     }
 
     private static Stream<Arguments> provideTestValuesForDistanceInMetres() {
       return Stream.of(
-          Arguments.of(0.0, 0.0, 0),
-          Arguments.of(10_000.0, 10.0, 60),
-          Arguments.of(10_000.0, 60.0, 10),
-          Arguments.of(0.0, 0.0, 7),
-          Arguments.of(0.0, 8.0, 0)
+          Arguments.of(0.0, 0, 0),
+          Arguments.of(10_000.0, 10, 60),
+          Arguments.of(10_000.0, 60, 10),
+          Arguments.of(0.0, 0, 7),
+          Arguments.of(0.0, 8, 0)
       );
     }
   }
@@ -60,7 +60,7 @@ class DistanceCalculatorBasedOnSpeedAndTimeTest {
     @ParameterizedTest
     @MethodSource("provideTestValues")
     void shouldCalculateDistance(double expected,
-                                 Double speedOverGroundInKmPerH,
+                                 Integer speedOverGroundInKmPerH,
                                  long minutesBetween) {
       //given
       //when
@@ -71,21 +71,21 @@ class DistanceCalculatorBasedOnSpeedAndTimeTest {
 
     @Test
     void shouldThrowErrorWhenSpeedIsNegative() {
-      assertThrows(IllegalStateException.class, () -> testObj.distanceInKilometres(-1.0, 1));
+      assertThrows(IllegalStateException.class, () -> testObj.distanceInKilometres(-1, 1));
     }
 
     @Test
     void shouldThrowErrorWhenTimeIsNegative() {
-      assertThrows(IllegalStateException.class, () -> testObj.distanceInKilometres(1.0, -1));
+      assertThrows(IllegalStateException.class, () -> testObj.distanceInKilometres(1, -1));
     }
 
     private static Stream<Arguments> provideTestValues() {
       return Stream.of(
-          Arguments.of(0.0, 0.0, 0),
-          Arguments.of(10.0, 10.0, 1),
-          Arguments.of(50.0, 10.0, 5),
-          Arguments.of(0.0, 0.0, 7),
-          Arguments.of(0.0, 8.0, 0)
+          Arguments.of(0.0, 0, 0),
+          Arguments.of(10.0, 10, 1),
+          Arguments.of(50.0, 10, 5),
+          Arguments.of(0.0, 0, 7),
+          Arguments.of(0.0, 8, 0)
       );
     }
   }
